facebook_id is BigInt

+++ apps/accounts/models.py
@@ -187,7 +187,7 @@ class RegistrationProfile(models.Model):
 	activation_key = models.CharField(_('activation key'), max_length=40)
 	reset_key = models.CharField(_('reset key'), max_length=40)
 	reset_key_created = models.DateTimeField(default=datetime.datetime.now())
-	facebook_id = models.IntegerField(null=True)
+	facebook_id = models.BigIntegerField(null=True)
 	quality_of_life = models.FloatField(null=True, max_length=6)
 	last_viewed_friends_shared = models.DateTimeField(default=datetime.datetime.now())
